refactor(python): sklearn base classes, ShapValues f32, types module
- Merge importance.rs and verbosity.rs into types.rs
- Change ShapValues storage from f64 to f32 for memory efficiency
  and consistency with prediction output
- Refactor sklearn estimators with shared base classes:
  - _GBDTEstimatorBase / _GBLinearEstimatorBase abstract base classes
  - Config created in __init__ with immediate validation
  - Add objective/metric parameters (accept enum types directly)
  - Add objective type validation (regressor vs classifier)
  - Add verbose parameter for sklearn API compatibility
  - eval_set accepts both tuple list and EvalSet list

All 548 Rust tests and 177 Python tests pass.

+++ crates/boosters/src/explainability/shap/tree_explainer.rs
@@ -305,9 +305,9 @@ mod tests {
         let shap = explainer.shap_values(view);
 
         // Sum SHAP values + base should equal prediction (-1.0)
-        let sum: f64 = (0..3).map(|f| shap.get(0, f, 0)).sum();
+        let sum: f32 = (0..3).map(|f| shap.get(0, f, 0)).sum();
         let prediction_from_shap = sum + shap.base_value(0, 0);
-        let _actual_prediction = -1.0; // leaf value
+        let _actual_prediction = -1.0f32; // leaf value
 
         // TODO: Verify SHAP algorithm correctness with reference values
         // The current implementation is a skeleton that needs validation
@@ -316,6 +316,6 @@ mod tests {
         
         // Feature 0 is the splitting feature, so it should have non-zero contribution
         let f0_contrib = shap.get(0, 0, 0);
-        assert!(f0_contrib != 0.0, "Splitting feature should have non-zero SHAP");
+        assert!(f0_contrib != 0.0f32, "Splitting feature should have non-zero SHAP");
     }
 }
